# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba tecnica - API de Transacciones",
        "description": "Documentación de ejemplo para Laravel 12 con Swagger",
        "contact": {
            "email": "soporte@midominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de pruebas"
        }
    ],
    "paths": {
        "/api/transacciones": {
            "get": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Listar todas las transacciones",
                "operationId": "824e64fbe148b905dc6e17d50e0b9389",
                "responses": {
                    "200": {
                        "description": "Listado de transacciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transacciones"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Crear una transacción",
                "operationId": "c59c0d1b3e08ed0e028c9aab2f774eea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-03 15:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transacción creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transacciones"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transacciones/{id}": {
            "get": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Obtener una transacción por ID",
                "operationId": "3dddb25c348f79d9e7bdc7f9cc28d7df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transacción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacción encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transacciones"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transacción no encontrada"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Listado de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un usuario",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "saldo_inicial": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "saldo_inicial": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Transacciones": {
                "title": "Transacciones",
                "required": [
                    "id",
                    "user_id",
                    "monto",
                    "fecha"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "monto": {
                        "type": "number",
                        "format": "float",
                        "example": 150.5
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-03 15:30:00"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "Usuario",
                "description": "Modelo de usuario de la aplicación",
                "required": [
                    "name",
                    "email",
                    "password",
                    "saldo_inicial"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@email.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "writeOnly": true,
                        "example": "secret123"
                    },
                    "saldo_inicial": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Transacciones",
            "description": "Transacciones"
        },
        {
            "name": "Usuarios",
            "description": "Usuarios"
        }
    ]
}